requirements.txt
crewai
crewai-tools
langchain-google-genai
python-dotenv
asyncio
browser-use
langchain-groq